+++ R/googleCloudStorageR.R
@@ -1,0 +1,7 @@
+#' googleCloudStorageR
+#'
+#' Interact with Google Cloud Storage API in R. Part of the 'cloudyr' project.
+#'
+#' @docType package
+#' @name googleCloudStorageR
+NULL